fix icon, text, constants
