--- src/components/AboutCiphers/AboutCiphers.constants.jsx
@@ -1,37 +1,0 @@
-import { nanoid } from 'nanoid';
-
-export const ciphersInfo = [
-  {
-    id: nanoid(),
-    name: 'AES',
-    description:
-      'Симметричное блочное шифрование, широко применяемое для защиты данных.',
-  },
-  {
-    id: nanoid(),
-    name: 'Код Цезаря',
-    description:
-      'Простой шифр сдвига, заменяет каждую букву другой с фиксированным сдвигом. Именован в честь Цезаря, римского императора.',
-  },
-  {
-    id: nanoid(),
-    name: 'Шифр Виженера',
-    description:
-      'Полиалфавитный шифр, использующий ключевое слово для замены букв. Предоставляет высокий уровень безопасности.',
-  },
-  {
-    id: nanoid(),
-    name: 'QR код',
-    description:
-      'Двухмерный штрих-код, содержащий различную информацию. Может быть отсканирован для доступа к закодированным данным.',
-  },
-  {
-    id: nanoid(),
-    name: 'Азбука Морзе',
-    description:
-      'Использует точки и тире для представления букв, цифр и знаков. Шифрует сообщения с помощью световых или звуковых сигналов.',
-  },
-  {
-    id: nanoid(),
-  },
-];
