+++ src/components/AboutCiphers/AboutCiphers.jsx
@@ -1,4 +1,3 @@
-import { ciphersInfo } from './AboutCiphers.constants';
 import styles from './AboutCiphers.module.scss';
 import Flipped from '../Flipped/Flipped';
 
@@ -6,7 +5,7 @@ const AboutCiphers = () => {
   return (
     <section className={styles.aboutCiphers} id="aboutCiphers">
       <h2 className={styles.aboutCiphers_title}>О шифрах</h2>
-       <Flipped ciphersInfo={ciphersInfo}/>
+      <Flipped />
     </section>
   );
 };
