+++ src/components/Flipped/Flipped.constants.jsx
@@ -1,0 +1,37 @@
+import { nanoid } from 'nanoid';
+
+export const ciphersInfo = [
+  {
+    id: nanoid(),
+    name: 'AES',
+    description:
+      'Advanced Encryption Standard — симметричный алгоритм блочного шифрования, который оперирует блоками по 128 бит. Создан в 1998 году. Сейчас алгоритм хорошо проанализирован и широко используется.',
+  },
+  {
+    id: nanoid(),
+    name: 'Код Цезаря',
+    description:
+      'Простой шифр сдвига, заменяет каждую букву другой с фиксированным сдвигом. Именован в честь римского императора Цезаря. Один из самых простых и наиболее широко известных методов шифрования.',
+  },
+  {
+    id: nanoid(),
+    name: 'Шифр Виженера',
+    description:
+      'Полиалфавитный шифр, использующий ключевое слово для замены букв. Этот метод является простой формой многоалфавитной замены. Шифр Виженера изобретался многократно. Впервые этот метод в своей книге описал Джовани Баттиста Белласо.',
+  },
+  {
+    id: nanoid(),
+    name: 'QR код',
+    description:
+      'Quick Response Code или Код быстрого реагирования —двухмерный штрих-код, предоставляющий информацию для быстрого ее распознавания с помощью камеры. Был изобретён в 1994 году в Японии и изначально разрабатывался для автомобильной промышленности.',
+  },
+  {
+    id: nanoid(),
+    name: 'Азбука Морзе',
+    description:
+      'Способ знакового кодирования, использующий короткие сигналы (точки) и длинные сигналы (тире) для представления букв, цифр и знаков. Шифрует сообщения с помощью световых или звуковых сигналов.',
+  },
+  {
+    id: nanoid(),
+  },
+];
