+++ src/components/Flipped/FlippedItem/FlippedItem.module.scss
@@ -3,14 +3,13 @@
 
 .flippedItem {
   list-style-type: none;
-  cursor: pointer;
   max-height: 360px;
   height: 100%;
   position: relative;
 
   &_item {
-    border-radius: 8px;
-    padding: 48px 40px;
+    border-radius: 20px;
+    padding: 48px 40px 60px 40px;
     width: 100%;
     height: 100%;
     overflow: hidden;
@@ -28,7 +27,7 @@
       border: none;
     }
     &_name {
-      margin-bottom: 10px;
+      margin-bottom: 20px;
       @include font-source-sans(28px, $white, 700, 32px);
     }
 
@@ -47,6 +46,8 @@
   &_svg {
     overflow: hidden;
     display: flex;
+    max-height: 180px;
+    max-width: 180px;
     & svg {
       width: 100%;
       height: 100%;
@@ -54,7 +55,7 @@
   }
 
   &_description {
-    @include font-source-sans(24px, $black, 400, 24px);
+    @include font-source-sans(18px, $black, 400, 24px);
   }
 
   &:hover > .flippedItem_front {
